+++ page/student/student_report.php
@@ -48,7 +48,7 @@
                             <td>$gender</td>
                             <td>{$row['dob']}</td>
                             <td>
-                                <a class='action-btn delete' href='student_delete.php?rowid={$row['id']}'>delete</a>
+                                <a class='action-btn delete' href='student_delete.php?rowid={$row['id']}'onclick='return confirmDelete()'>delete</a>
                                 <a class='action-btn update' href='student_update.php?rowid={$row['id']}'>update</a>
                             </td>
 
@@ -61,7 +61,11 @@
     <br>
 
     
-
+    <script>
+        function confirmDelete() {
+            return confirm("Are you sure you want to delete this record?");
+        }
+    </script>
 </body>
 
 </html>
